% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{point_extract}
\alias{point_extract}
\title{point_extract}
\usage{
point_extract(data, gpkg)
}
\arguments{
\item{data}{A terra SpatRaster object}

\item{gpkg}{An sf object}
}
\value{
A dataframe with point values for the geopackage points
}
\description{
Extract point values for a raster
}
\examples{
result <- point_extract(some_raster,some_gpkg)

}
