% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{run_calc}
\alias{run_calc}
\title{run_calc}
\usage{
run_calc(bands, calc, data)
}
\arguments{
\item{bands}{A vector or list with the raster's band order}

\item{calc}{A string containing the calculation to be run. Variables in this
calculation are represented like so:
b - blue band
g - green band
r - red band
n - nir band
re - rededge band}

\item{data}{A terra SpatRaster object}
}
\value{
A terra SpatRaster object with the resulting raster after the
     calculation has been performed
}
\description{
Run an index calculation on an raster
}
\examples{
result <- run_calc(c("red","green","blue"),"b/g",some_raster)

}
