% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{run_indices}
\alias{run_indices}
\title{run_indices}
\usage{
run_indices(
  to_run,
  indices,
  bands,
  img_dir,
  proc_dir,
  gpkg,
  save_dir = "",
  points = FALSE
)
}
\arguments{
\item{to_run}{A vector or list of index names to be run}

\item{indices}{A named list generator from the read_indices function}

\item{bands}{A vector or list of ordered band names for the images}

\item{img_dir}{File path to folder containing images to run on}

\item{proc_dir}{File path to the processing folder for temporary files}

\item{gpkg}{File path to a geopackage or an sf object}
}
\value{
An sf dataframe with geopackage data and extracted index information
}
\description{
Run multiple vegetation index calculation over multiple images,
     and extract their values per polygon or point as provided in geopackage.
}
\examples{
result <- point_extract(some_raster,some_gpkg)

}
