% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{volume_calc}
\alias{volume_calc}
\title{volume_calc}
\usage{
volume_calc(gpkg, data)
}
\arguments{
\item{gpkg}{File path to a geopackage or an sf object}

\item{data}{File path to a DSM/DEM tif or a terra SpatRaster object}
}
\value{
A terra SpatRaster object containing volumes for each pixel, to be
   extracted using exact_extract
}
\description{
Calculate volume for each polygon over a DSM/DEM raster using plane averages
}
\examples{
volumes <- volume_calc("/path/to/polygons.gpkg","/path/to/dsm.tif")

}
