% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{zonal_stats}
\alias{zonal_stats}
\title{zonal_stats}
\usage{
zonal_stats(
  img_dir,
  bands,
  to_run,
  gpkg,
  out_gpkg = "",
  save_dir = "",
  points = FALSE,
  buffer = 0
)
}
\arguments{
\item{img_dir}{File path to folder with rasters}

\item{bands}{List or vector containing band names in order for rasters}

\item{to_run}{List or vector containing the index names to run.
Volume can also be calculated using index VOLUME.
Raw extraction of data can be performed using index RAW.}

\item{gpkg}{File path to geopackage or sf object}

\item{out_gpkg}{Optional, path to an output geopackage}

\item{save_dir}{Optional, path to a folder to save calculation rasters in}

\item{points}{Set to FALSE by default, set to TRUE if geopackage uses points}

\item{buffer}{Set to 0 be default, set to a nonzero number to
generate buffers around points. When using buffer, please have your
geopackage and raster in a CRS which uses meters}
}
\value{
An sf dataframe with geopackage data and extracted index information
}
\description{
Perform index or volume calculations, and extract data by
     polygon or points.
}
